change swagger-io dependency to springfox dependecy. It was adjusted unit class.

+++ src/main/java/com/demo/currencyconverter/dto/CurrencyRateDTO.java
@@ -3,7 +3,7 @@ package com.demo.currencyconverter.dto;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.util.List;
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
@@ -11,6 +11,6 @@ public class CurrencyRateDTO implements Serializable {
 
 	private String success;
 	private String base;
-	private Map<String,Double> rates;
+	private Map<String, BigDecimal> rates;
 	
 }
