+++ src/main/java/com/demo/currencyconverter/dto/RateDTO.java
@@ -5,11 +5,12 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class RateDTO implements Serializable {
 
-	private Double symbols;
+	private BigDecimal symbols;
 }
